Develop

+++ src/TD.Komercijalno/TD.Komercijalno.Contracts/Dtos/VrstaDoks/VrstaDokDto.cs
@@ -2,7 +2,7 @@
 {
     public class VrstaDokDto
     {
-        public int VrDok { get; set; }
+        public long VrDok { get; set; }
         public string NazivDok { get; set; }
         public int? Poslednji { get; set; }
         public short? Io { get; set; }
